requirements.txt
openai
langchain
faiss-cpu
gradio
python-dotenv
